update: move to v0.2.0
