--- src/laana-v0.1.3/proto/entity/file.proto
@@ -1,10 +1,0 @@
-syntax = 'proto3';
-package Laana;
-
-message File {
-  oneof uri {
-    string url = 1;
-    string cacheId = 2;
-    bytes raw = 3;
-  }
-}
