+++ src/laana-v0.2.0/action/index.ts
@@ -1,4 +1,4 @@
-import { ActionPing, ActionPong } from '@/laana-v0.1.3/types/action/wrapper';
+import { ActionPing, ActionPong } from '../types/action/wrapper';
 
 type ExtractFromPongOrVoid<key> = Extract<ActionPong['pong'], { oneofKind: key; }> extends never ?
     void :
