+++ src/laana-v0.2.0/index.ts
@@ -1,12 +1,12 @@
 import { InstanceContext, NapCatCore } from '@/core';
 import { NapCatPathWrapper } from '@/common/path';
-import { LaanaFileUtils } from '@/laana-v0.1.3/utils/file';
-import { LaanaMessageUtils } from '@/laana-v0.1.3/utils/message';
-import { LaanaActionHandler } from '@/laana-v0.1.3/action';
-import { LaanaMessageActionHandler } from '@/laana-v0.1.3/action/message';
-import { LaanaConfigLoader } from '@/laana-v0.1.3/config';
-import { LaanaNetworkManager } from '@/laana-v0.1.3/network';
-import { LaanaWsServerAdapter } from '@/laana-v0.1.3/network/ws-server';
+import { LaanaFileUtils } from './utils/file';
+import { LaanaMessageUtils } from './utils/message';
+import { LaanaActionHandler } from './action';
+import { LaanaMessageActionHandler } from './action/message';
+import { LaanaConfigLoader } from './config';
+import { LaanaNetworkManager } from './network';
+import { LaanaWsServerAdapter } from './network/ws-server';
 
 export class NapCatLaanaAdapter {
     utils = {
