+++ src/laana-v0.2.0/proto/event/buddy.proto
